fix mob:hurt

+++ LiteLoader/Header/MC/Actor.hpp
@@ -50,7 +50,7 @@ public:
     LIAPI bool setOnFire(int time, bool isEffect);
     LIAPI bool stopFire();
     LIAPI bool hasTag(const string& tag);
-	LIAPI bool hurtEntity(int damage);
+	LIAPI bool hurtEntity(float damage);
     LIAPI bool teleport(Vec3 to, int dimID, float x, float y);
 	LIAPI bool teleport(Vec3 pos,int dimid);
     LIAPI ItemStack* getHandSlot();
